+++ src/main/java/edu/isgb/school/repositories/SchoolRepository.java
@@ -1,7 +1,7 @@
 package edu.isgb.school.repositories;
 
-import edu.isgb.school.entities.School;
 import org.springframework.data.jpa.repository.JpaRepository;
+import edu.isgb.school.entities.School;
 
-public interface SchoolRepository extends JpaRepository<School, Long> {
+public interface SchoolRepository extends JpaRepository<School, Integer> {
 }
